+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Install container registry
+- Install kpack
+- Install kpack cluster stores and cluster stacks
 
 ### Changed